update file extraction

+++ data/river_flow.py
@@ -1,6 +1,7 @@
+import os
 import zipfile
-import urllib.request
 import pandas as pd
+import urllib.request
 
 
 def download_data():
@@ -9,10 +10,13 @@ def download_data():
 
         Side effect: Creating file `amax.csv`.
     '''
-    url = 'https://nrfa.ceh.ac.uk/sites/default/files/NRFAPeakFlow_v10.zip'
-    urllib.request.urlretrieve(url, 'NRFAPeakFlow_v10.zip')
-    with zipfile.ZipFile('NRFAPeakFlow_v10.zip', 'r') as zip_ref:
-        zip_ref.extract('amax.csv', path='./downloaded_NRFA')
+    if not os.path.exists('data/NRFAPeakFlow_v10.zip'):
+        url = 'https://nrfa.ceh.ac.uk/sites/default/files/NRFAPeakFlow_v10.zip'
+        urllib.request.urlretrieve(url, 'NRFAPeakFlow_v10.zip')
+    
+    if not os.path.exists('downloaded_NRFA/amax.csv'):    
+        with zipfile.ZipFile('NRFAPeakFlow_v10.zip', 'r') as zip_ref:
+            zip_ref.extract('amax.csv', path='./downloaded_NRFA')
 
 
 def extract(datafile_name):
@@ -22,14 +26,15 @@ def extract(datafile_name):
         Side effect: `"./data/{river}.csv"` files created, 
                     each contains a series of peak flow rates sorted descending.
     '''
-    df = pd.read_csv('./downloaded_NRFA/amax.csv')
-    river_ids = pd.read_csv('rivers.csv')['StationNo'].to_list()
+    df = pd.read_csv(datafile_name)
+    river_ids = pd.read_csv('data/river.csv')['StationNo'].to_list()
     for river in river_ids:
         if not river.startswith('?'):
-            river = '0' + river
-        df = df[df['StationNo'] == river]
-        df.to_csv(f'./data/{river}.csv', index=False)
+            df = df[df['river'] == river]
+            df.to_csv(f'./data/{river}.csv', index=False)
+
 
 if __name__ == '__main__':
-    download_data()
+    # download_data()
+    extract('./downloaded_NRFA/amax.csv')
     print('Done')
